Refactor CLI code to follow MVC. this involved creating a controller class and two helper classes display and exists to adhere to the data access restrictions of MVC. Other modified files are main and menu which have been refactored to have main just start the program and menu to handle all user input as the VIEW
