+++ classes.py
@@ -4,32 +4,28 @@
 from diagram import diagram
 
 def addClass(name):
-    #checks if name is already in the diagram to avoid repeats
+    #checks if name is already in the diagram to avoid repeats returns False
     if name in diagram:
-        print(f"'{name}' already found in the diagram.")
+        return False
     
-    #adds the class if it is not in the class already
+    #adds the class if it is not in the class already returns True
     else:
         diagram.update({name: {}})
-        print(f"'{name}' added to the diagram.")
+        return True
 
 def renameClass(oldName, newName):
-    #checks if the old name is in the diagram then updates the name if it is
+    #checks if the old name is in the diagram then updates to new name
     if oldName in diagram:
         diagram[newName] = diagram.pop(oldName)
-        print(f"'{oldName}' has been renamed to '{newName}'")
-    
-    #if old name is not in the diagram it gives an error message
+        return True
     else:
-        print(f"'{oldName}' not found in the diagram.")
+        return False
         
 def deleteClass(name):
-    #checks if name is in the diagram then deletes it if it is
+    #checks if name is in the diagram then deletes it
     if name in diagram:
         del diagram[name]
-        print(f"'{name}' removed from the diagram.")
-    
-    #if name is not in the diagram it returns an error message
+        return True
     else:
-        print(f"'{name}' not found in the diagram.")
+        return False
